+++ .rake_tasks~
@@ -1,0 +1,67 @@
+about
+action_mailbox:ingress:exim
+action_mailbox:ingress:postfix
+action_mailbox:ingress:qmail
+action_mailbox:install
+action_text:install
+active_storage:install
+app:template
+app:update
+assets:clean[keep]
+assets:clobber
+assets:environment
+assets:precompile
+cache_digests:dependencies
+cache_digests:nested_dependencies
+db:create
+db:drop
+db:environment:set
+db:fixtures:load
+db:migrate
+db:migrate:status
+db:prepare
+db:rollback
+db:schema:cache:clear
+db:schema:cache:dump
+db:schema:dump
+db:schema:load
+db:seed
+db:seed:replant
+db:setup
+db:structure:dump
+db:structure:load
+db:version
+log:clear
+middleware
+restart
+secret
+stats
+test
+test:db
+test:system
+time:zones[country_or_offset]
+tmp:clear
+tmp:create
+webpacker
+webpacker:binstubs
+webpacker:check_binstubs
+webpacker:check_node
+webpacker:check_yarn
+webpacker:clean[keep]
+webpacker:clobber
+webpacker:compile
+webpacker:info
+webpacker:install
+webpacker:install:angular
+webpacker:install:coffee
+webpacker:install:elm
+webpacker:install:erb
+webpacker:install:react
+webpacker:install:stimulus
+webpacker:install:svelte
+webpacker:install:typescript
+webpacker:install:vue
+webpacker:verify_install
+webpacker:yarn_install
+yarn:install
+zeitwerk:check
