+++ app/jobs/daily_report_job.rb
@@ -1,0 +1,7 @@
+class DailyReportJob < ApplicationJob
+  queue_as :default
+
+  def perform
+    SessionsController.new.daily_report
+  end
+end
