+++ app/workers/daily_report_worker.rb
@@ -1,0 +1,7 @@
+class DailyReportWorker
+  include Sidekiq::Worker
+
+  def perform
+    # Do something
+  end
+end
