+++ config/application.rb
@@ -13,6 +13,7 @@ module NewApp
     # Initialize configuration defaults for originally generated Rails version.
     config.load_defaults 6.0
     config.action_view.embed_authenticity_token_in_remote_forms = true
+    config.active_job.queue_adapter = :sidekiq
     # Settings in config/environments/* take precedence over those specified here.
     # Application configuration can go into files in config/initializers
     # -- all .rb files in that directory are automatically loaded after loading
