+++ config/routes.rb
@@ -1,5 +1,7 @@
 # frozen_string_literal: true
 
+require "sidekiq/web"
+
 Rails.application.routes.draw do
   devise_for :users, controllers: {
     registrations: "users/registrations",
@@ -32,6 +34,8 @@ Rails.application.routes.draw do
   end
   resources :relationships, only: %i(create destroy)
   mount ActionCable.server => "/cable"
+  mount Sidekiq::Web => "/sidekiq"
+
   # For details on the DSL available within this file, see https://guides.rubyonrails.org/routing.html
   root "static_pages#home"
 end
